Fold tests/test-https-strict.js into tests/test-https.js
